Move generic predicates from `extensions/pkg/predicate` to `pkg/controllerutils/predicate`
