+++ extensions/pkg/controller/bastion/controller.go
@@ -23,6 +23,7 @@ import (
 
 	extensionspredicate "github.com/gardener/gardener/extensions/pkg/predicate"
 	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
+	predicateutils "github.com/gardener/gardener/pkg/controllerutils/predicate"
 )
 
 const (
@@ -57,7 +58,7 @@ func DefaultPredicates(ignoreOperationAnnotation bool) []predicate.Predicate {
 
 	return []predicate.Predicate{
 		predicate.Or(
-			extensionspredicate.HasOperationAnnotation(),
+			predicateutils.HasOperationAnnotation(),
 			extensionspredicate.LastOperationNotSuccessful(),
 			extensionspredicate.IsDeleting(),
 		),
