+++ extensions/pkg/controller/extension/reconciler.go
@@ -19,102 +19,23 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/go-logr/logr"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-	"sigs.k8s.io/controller-runtime/pkg/handler"
-	"sigs.k8s.io/controller-runtime/pkg/log"
-	"sigs.k8s.io/controller-runtime/pkg/manager"
-	"sigs.k8s.io/controller-runtime/pkg/predicate"
-	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-	"sigs.k8s.io/controller-runtime/pkg/runtime/inject"
-	"sigs.k8s.io/controller-runtime/pkg/source"
-
 	extensionscontroller "github.com/gardener/gardener/extensions/pkg/controller"
 	"github.com/gardener/gardener/extensions/pkg/controller/common"
-	extensionspredicate "github.com/gardener/gardener/extensions/pkg/predicate"
 	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	v1beta1constants "github.com/gardener/gardener/pkg/apis/core/v1beta1/constants"
 	gardencorev1beta1helper "github.com/gardener/gardener/pkg/apis/core/v1beta1/helper"
 	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
 	"github.com/gardener/gardener/pkg/controllerutils"
-	"github.com/gardener/gardener/pkg/controllerutils/mapper"
 	kutil "github.com/gardener/gardener/pkg/utils/kubernetes"
+
+	"github.com/go-logr/logr"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+	"sigs.k8s.io/controller-runtime/pkg/log"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+	"sigs.k8s.io/controller-runtime/pkg/runtime/inject"
 )
-
-const (
-	// FinalizerPrefix is the prefix name of the finalizer written by this controller.
-	FinalizerPrefix = "extensions.gardener.cloud"
-)
-
-// AddArgs are arguments for adding an Extension resources controller to a manager.
-type AddArgs struct {
-	// Actuator is an Extension resource actuator.
-	Actuator Actuator
-	// Name is the name of the controller.
-	Name string
-	// FinalizerSuffix is the suffix for the finalizer name.
-	FinalizerSuffix string
-	// ControllerOptions are the controller options used for creating a controller.
-	// The options.Reconciler is always overridden with a reconciler created from the
-	// given actuator.
-	ControllerOptions controller.Options
-	// Predicates are the predicates to use.
-	Predicates []predicate.Predicate
-	// Resync determines the requeue interval.
-	Resync time.Duration
-	// Type is the type of the resource considered for reconciliation.
-	Type string
-	// IgnoreOperationAnnotation specifies whether to ignore the operation annotation or not.
-	// If the annotation is not ignored, the extension controller will only reconcile
-	// with a present operation annotation typically set during a reconcile (e.g in the maintenance time) by the Gardenlet
-	IgnoreOperationAnnotation bool
-}
-
-// Add adds an Extension controller to the given manager using the given AddArgs.
-func Add(mgr manager.Manager, args AddArgs) error {
-	args.ControllerOptions.Reconciler = NewReconciler(args)
-	return add(mgr, args)
-}
-
-// DefaultPredicates returns the default predicates for an extension reconciler.
-func DefaultPredicates(ignoreOperationAnnotation bool) []predicate.Predicate {
-	if ignoreOperationAnnotation {
-		return []predicate.Predicate{
-			predicate.GenerationChangedPredicate{},
-		}
-	}
-	return []predicate.Predicate{
-		predicate.Or(
-			extensionspredicate.HasOperationAnnotation(),
-			extensionspredicate.LastOperationNotSuccessful(),
-			extensionspredicate.IsDeleting(),
-		),
-		extensionspredicate.ShootNotFailed(),
-	}
-}
-
-func add(mgr manager.Manager, args AddArgs) error {
-	ctrl, err := controller.New(args.Name, mgr, args.ControllerOptions)
-	if err != nil {
-		return err
-	}
-
-	predicates := extensionspredicate.AddTypePredicate(args.Predicates, args.Type)
-
-	if args.IgnoreOperationAnnotation {
-		if err := ctrl.Watch(
-			&source.Kind{Type: &extensionsv1alpha1.Cluster{}},
-			mapper.EnqueueRequestsFrom(ClusterToExtensionMapper(predicates...), mapper.UpdateWithNew),
-		); err != nil {
-			return err
-		}
-	}
-
-	return ctrl.Watch(&source.Kind{Type: &extensionsv1alpha1.Extension{}}, &handler.EnqueueRequestForObject{}, predicates...)
-}
 
 // reconciler reconciles Extension resources of Gardener's
 // `extensions.gardener.cloud` API group.
